requirements.txt
PyYAML
future
mock ; python_version < '3.3'
unittest2 ; python_version < '3'

